+++ src/harness_evals/metrics/llm_judge/__init__.py
@@ -13,10 +13,12 @@ from harness_evals.metrics.llm_judge.pairwise import PairwiseMetric
 from harness_evals.metrics.llm_judge.prompt_alignment import PromptAlignmentMetric
 from harness_evals.metrics.llm_judge.rubric_judge import RubricJudgeMetric
 from harness_evals.metrics.llm_judge.summarization import SummarizationMetric
+from harness_evals.metrics.llm_judge.types import RubricLevel
 
 __all__ = [
     "GEvalMetric",
     "RubricJudgeMetric",
+    "RubricLevel",
     "PairwiseMetric",
     "DAGMetric",
     "DeepAcyclicGraph",